+++ docs/background/releases.rst
@@ -9,6 +9,10 @@ Release History
 * Removed automatic conversion of ``NSString`` objects to ``str`` when returned from Objective-C methods. This feature made it difficult to call Objective-C methods on ``NSString`` objects, because there was no easy way to prevent the automatic conversion.
   * In most cases, this change will not affect existing code, because ``NSString`` objects now support operations similar to ``str``.
   * If an actual ``str`` object is required, the ``NSString`` object can be wrapped in a ``str`` call to convert it.
+* Added support for ``objc_property``s with non-object types.
+* Added public ``get_ivar`` and ``set_ivar`` functions for manipulating ivars.
+* Changed the implementation of ``objc_property`` to use ivars instead of Python attributes for storage. This fixes name conflicts in some situations.
+* Fixed ``objc_property`` setters on non-macOS platforms. (cculianu)
 * Fixed various bugs in the collection ``ObjCInstance`` subclasses:
   * Fixed getting/setting/deleting items or slices with indices lower than ``-len(obj)``. Previously this crashed Python, now an ``IndexError`` is raised.
   * Fixed slices with step size 0. Previously they were ignored and 1 was incorrectly used as the step size, now an ``IndexError`` is raised.